+++ src/ansys/dyna/core/keywords/keyword_classes/auto/include.py
@@ -20,12 +20,11 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import typing
-from ansys.dyna.core.lib.cards_.special.include_card import IncludeCard
+from ansys.dyna.core.lib.cards_.special.include_card import IncludeCard, IncludeCardMixin
 from ansys.dyna.core.lib.keyword_base import KeywordBase
 
 
-class Include(KeywordBase):
+class Include(KeywordBase, IncludeCardMixin):
     """DYNA INCLUDE keyword"""
 
     keyword = "INCLUDE"
@@ -33,16 +32,4 @@ class Include(KeywordBase):
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
-        self._cards = [IncludeCard(kwargs.get("filename"))]
-
-    @property
-    def filename(self) -> typing.Optional[str]:
-        """Get or set the File name of file to be included in this keyword file.
-        Maximum 80 charcters. If the STAMPED_PART option is active, this is the DYNAIN file containing the results from metal stamping.
-        """ # nopep8
-        return self._cards[0].get_value("filename")
-
-    @filename.setter
-    def filename(self, value: str) -> None:
-        self._cards[0].set_value("filename", value)
-
+        self._cards = [IncludeCard(**kwargs)]
